#11-feat: document all routes in app
