+++ api/src/routes/index.ts
@@ -1,3 +1,4 @@
+import { Express } from 'express';
 import { Router } from 'express';
 import appRoutes from './app.routes';
 import userRoutes from './user.routes';
@@ -5,4 +6,12 @@ import authRoutes from './auth.routes';
 
 const routes: Router[] = [appRoutes, userRoutes, authRoutes];
 
-export default routes;
+export default function initRoutes(app: Express) {
+  try {
+    routes.forEach((route) => {
+      app.use(route);
+    });
+  } catch (error: any) {
+    console.log('❗ Routes could not be initialized', error);
+  }
+}
