--- api/src/routes/init.routes.ts
@@ -1,12 +1,0 @@
-import { Express } from 'express';
-import routes from './';
-
-export default function initRoutes(app: Express) {
-  try {
-    routes.forEach((route) => {
-      app.use(route);
-    });
-  } catch (error: any) {
-    console.log('❗ Routes could not be initialized', error);
-  }
-}
